divide twitter login and get tweets
Divide TwitterActivityControll to TwitterLoginController and TwitterTweetController.
That is because I make controllers depend on function, not activity.

In TwitterActivity, remove two buttons. It only needs a twitter login button.
And check token_type and access_token for auto login.

In MapsActivity, add SeekBar widget to change radius, and TwitterTweetController
to get tweets.

Add BaseActivity for capsulation.